Refactor code: Extract components and create custom hooks
