+++ src/hooks/useScroll/useScroll.jsx
@@ -1,0 +1,18 @@
+import { useState, useEffect } from "react";
+
+export const useScroll = () => {
+  const [showScrollButton, setShowScrollButton] = useState(false);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      setShowScrollButton(window.pageYOffset > 0);
+    };
+
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, []);
+
+  return showScrollButton;
+};
